Fixes link color change issues.
It turns out that there are two LinkIt Links for each MySystem Link - one controlled by the startNode and one by the endNode. To fully change the link label/color, we need to notify both nodes of the change. That has to happen on a change, not on system load, or the system goes into infinite loop.