add itemColorKey, fix input element values
This includes a hack because there isn't time to figure out what is really going on.  As the comment says it isn't clear why there is value property along with a getFieldValue method.  It appears the class was taking from SproutCore and modified, but there is also a patched_form_radio class that might be at work here.