Handle case with no energy types at all, and don't show inspector on selection either.
Inspector will now not show at all if energy types < 2. This may confuse authors who expect it to show
when energy type == 1. If so, it can easily be changed without any side-effects (modal dialog would
still not show).

Also, in the future we may want the inspector to show other information, so this would have to be
reinstated.