test to verify hudson is loading test out of order
The order the test are loaded should not matter.  However the fact that the build environments between the CI server and our local machines makes tracking down problems difficult.  Making the order be deterministic will prevent these difficult to find problems.  It will also mean ordering issues will be not be found.  However if ordering problems are a problem then there could be a version of this that randomizes the order of the tests.