Calculate arrow end directly from mouse pos + view offset, instead of using delta
Using deltaX and deltaY, the arrow was not in the right place if the user
started from anywhere other than the dead-center of the node.

By just using the mouse position directly (plus the view offset to account for
the difference between the view's coords and the screen coords), we can always
draw the arrow right at the mouse position.