only SC initiated save sends learner data over net
Previously even a Wise4 initiated save would
send the learner data over the network.  This was 
unnecessary because Wise4 sends the data anyhow when
it initiates the save.  Wise4 initiates a save by calling
Mysystem2Node#onExit, which in turn calls window.save,
which then calls the Mysystem2#save.

This commit also adds an integration testing framework
for the step.