Part of [#32532217] - Author-able background image.
https://www.pivotaltracker.com/story/show/32532217

An author of a step in an activity can place an image on the background of MySystem to help scaffold the task of making a diagram.
The image should scale automatically to the size of the window containing the diagram. We will use the centered letterboxing approach to manage mismatching proportions.